Rename airbyte-metrics module to airbyte-metrics-lib.
