+++ airbyte-metrics/lib/src/main/java/io/airbyte/metrics/lib/DogstatsdMetricSingleton.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2021 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.metrics;
+package io.airbyte.metrics.lib;
 
 import com.timgroup.statsd.NonBlockingStatsDClientBuilder;
 import com.timgroup.statsd.StatsDClient;
