+++ airbyte-scheduler/app/build.gradle
@@ -12,7 +12,7 @@ dependencies {
     implementation project(':airbyte-config:persistence')
     implementation project(':airbyte-db:lib')
     implementation project(':airbyte-json-validation')
-    implementation project(':airbyte-metrics')
+    implementation project(':airbyte-metrics:lib')
     implementation project(':airbyte-protocol:models')
     implementation project(':airbyte-scheduler:models')
     implementation project(':airbyte-scheduler:persistence')
